Module Flux.Stream

val run : from:'a source -> via:('a, 'b) flow -> into:('b, 'c) sink -> 'c * 'a source option

Fuses sources, sinks and flows and produces a result and a leftover.

  let r, leftover = Stream.run ~from:source ~via:flow ~into:sink

Streams elements from source into sink via a stream transformer flow. In addition to the result value r produced by sink, a leftover source is returned, if source was not exhausted.

NOTE: If a leftover source is produced, it is required to either consume it or manually dispose its resources. Not doing so might lead to resource leaks.

val into : ('a, 'b) sink -> 'a stream -> 'b

into sink stream is the result value produced by streaming all elements of stream into sink.

val via : ('a, 'b) flow -> 'a stream -> 'b stream

via flow stream is stream produced by transforming all elements of stream via flow.

val from : 'a source -> 'a stream

from source is a stream created from a source.

val map : ('a -> 'b) -> 'a stream -> 'b stream

A stream with all elements transformed with a mapping function.

val flat_map : ('a -> 'b stream) -> 'a stream -> 'b stream

flat_map fn stream is a stream concatenated from sub-streams produced by applying fn to all elements of stream.

val filter : ('a -> bool) -> 'a stream -> 'a stream

A stream that includes only the elements that satisfy a predicate.

val file : filename:string -> string stream -> unit

file filename stream writes bytes from stream into the file located at filename.

val drain : 'a stream -> unit

Consume and discard all elements from the given stream.

val interpose : 'a -> 'a stream -> 'a stream

Inserts a separator element between each stream element.

val each : ?parallel:bool -> ('a -> unit) -> 'a stream -> unit

each fn stream applies an function fn to all elements of stream. Each function is performed cooperatively via the Miou scheduler.

val empty : 'a stream

Empty stream with no elements.

val range : ?by:int -> int -> int -> int stream

range ~by:step n m is a sequence of integers starting from n to m (excluding m) incremented by step. The range is pen on the right side.

val repeat : ?times:int -> 'a -> 'a stream

repeat ~times:n x produces a stream by repeating x n times. If n is omitted, x is repeated ad infinitum.